+++ src/eo_processing/resources/udf_catboost_inference.py
@@ -1,0 +1,128 @@
+import functools
+import sys
+from typing import Dict
+import xarray as xr
+import numpy as np
+
+# The onnx_deps folder contains the extracted contents of the dependencies archive provided in the job options
+sys.path.insert(0, "onnx_deps") 
+import onnxruntime as ort
+
+@functools.lru_cache(maxsize=5)
+def load_onnx_model(model_name: str) -> ort.InferenceSession:
+    """
+    Loads an ONNX model from the onnx_models folder and returns an ONNX runtime session.
+
+    Extracting the model loading code into a separate function allows us to cache the loaded model.
+    This prevents the model from being loaded for every chunk of data that is processed, but only once per executor,
+    which can save a lot of time, memory and ultimately processing costs.
+
+    Should you have to download the model from a remote location, you can add the download code here, and cache the model.
+
+    Make sure that the arguments of the method you add the @functools.lru_cache decorator to are hashable.
+    Be careful with using this decorator for class methods, as the self argument is not hashable. 
+    In that case you can use a static method or make sure your class is hashable (more difficult): https://docs.python.org/3/faq/programming.html#faq-cache-method-calls.
+
+    More information on this functool can be found here: 
+    https://docs.python.org/3/library/functools.html#functools.lru_cache
+    """
+    # The onnx_models folder contains the content of the model archive provided in the job options
+    return ort.InferenceSession(f"onnx_models/{model_name}")
+
+def preprocess_input(input_xr: xr.DataArray, ort_session: ort.InferenceSession) -> tuple:
+    """
+    Preprocess the input DataArray by ensuring the dimensions are in the correct order,
+    reshaping it, and returning the reshaped numpy array and the original shape.
+    """
+    input_xr = input_xr.transpose("y", "x", "bands")
+    input_shape = input_xr.shape
+    input_np = input_xr.values.reshape(-1, ort_session.get_inputs()[0].shape[1])
+    return input_np, input_shape
+
+def run_inference(input_np: np.ndarray, ort_session: ort.InferenceSession) -> tuple:
+    """
+    Run inference using the ONNX runtime session and return predicted labels and probabilities.
+    """
+    ort_inputs = {ort_session.get_inputs()[0].name: input_np}
+    ort_outputs = ort_session.run(None, ort_inputs)
+    predicted_labels = ort_outputs[0]
+    probabilities_dicts = ort_outputs[1]
+    return predicted_labels, probabilities_dicts
+
+def postprocess_output(predicted_labels: np.ndarray, probabilities_dicts: list, input_shape: tuple) -> tuple:
+    """
+    Postprocess the output by reshaping the predicted labels and probabilities into the original spatial structure.
+    """
+    class_labels = list(probabilities_dicts[0].keys())
+
+    # Convert probabilities into a 2D array
+    probabilities = np.array([[prob[class_id] for class_id in class_labels] for prob in probabilities_dicts])
+
+    # Reshape to match the (y, x) spatial structure
+    predicted_labels = predicted_labels.reshape(input_shape[0], input_shape[1])
+    probabilities = probabilities.reshape(len(class_labels), input_shape[0], input_shape[1])
+
+    return predicted_labels, probabilities
+
+def create_output_xarray(predicted_labels: np.ndarray, probabilities: np.ndarray, 
+                         input_xr: xr.DataArray) -> xr.DataArray:
+    """
+    Create an xarray DataArray with predicted labels and probabilities stacked along the bands dimension.
+    """
+    combined_data = np.concatenate([
+        predicted_labels[np.newaxis, :, :],  # Shape (1, y, x) for predicted labels
+        probabilities  # Shape (n_classes, y, x) for probabilities
+    ], axis=0)
+
+    return xr.DataArray(
+        combined_data,
+        dims=["bands", "y", "x"],
+        coords={
+            'y': input_xr.coords['y'],
+            'x': input_xr.coords['x']
+        }
+    )
+
+def apply_model(input_xr: xr.DataArray) -> xr.DataArray:
+    """
+    Run inference on the given input data using the provided ONNX runtime session.
+    This method is called for each timestep in the chunk received by apply_datacube.
+    """
+    # Step 1: Load the ONNX model
+    ort_session = load_onnx_model("test.onnx")
+
+    # Step 2: Preprocess the input
+    input_np, input_shape = preprocess_input(input_xr, ort_session)
+
+    # Step 3: Perform inference
+    predicted_labels, probabilities_dicts = run_inference(input_np, ort_session)
+
+    # Step 4: Postprocess the output
+    predicted_labels, probabilities = postprocess_output(predicted_labels, probabilities_dicts, input_shape)
+
+    # Step 5: Create the output xarray
+    return create_output_xarray(predicted_labels, probabilities, input_xr)
+
+def apply_datacube(cube: xr.DataArray, context: Dict) -> xr.DataArray:
+    """
+    Function that is called for each chunk of data that is processed.
+    The function name and arguments are defined by the UDF API.
+    
+    More information can be found here: 
+    https://open-eo.github.io/openeo-python-client/udf.html#udf-function-names-and-signatures
+
+    CAVEAT: Some users tend to extract the underlying numpy array and preprocess it for the model using Numpy functions.
+        The order of the dimensions in the numpy array might not be the same for each back-end or when running a udf locally, 
+        which can lead to unexpected results. 
+
+        It is recommended to use the named dimensions of the xarray DataArray to avoid this issue.
+        The order of the dimensions can be changed using the transpose method.
+        While it is a better practice to do preprocessing using openeo processes, most operations are also available in Xarray. 
+    """
+    # Define how you want to handle nan values
+    cube = cube.fillna(0)
+
+    # Apply the model for each timestep in the chunk
+    output_data = apply_model(cube)
+
+    return output_data
