+++ src/eo_processing/utils/__init__.py
@@ -1,6 +1,6 @@
 from .geoprocessing import (laea20km_id_to_extent,
                             reproj_bbox_to_ll,
                             bbox_area)
-from .helper import init_connection, location_visu
+from .helper import init_connection, location_visu, getUDFpath
 
-__all__ = ['laea20km_id_to_extent', 'reproj_bbox_to_ll', 'bbox_area', 'init_connection', 'location_visu']
+__all__ = ['laea20km_id_to_extent', 'reproj_bbox_to_ll', 'bbox_area', 'init_connection', 'location_visu', 'getUDFpath']
